Fix Function Test for hdfsConcat() Success Case
There was an incorrect file name in the hdfsConcat() success
test case, but, when I corrected that, I discovered that the
underlying HDFS on my test platform required that the test
pre-create the concatenation target (output) file before calling
hdfsConcat().  Since I don't care what the underlying HDFS
requires as long as calling hdfsConcat() ultimately has the
desired effect, I beefed up the test to verify correct data in
the target file and added the pre-creation (which is harmless
on any underlying HDFS implementation).